# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hoja de Ruta - Gas/Garrafa delivery route management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Cliente {
  id: number;
  nombre: string;
  /** @nullable */
  dir: string | null;
}

export interface ClienteInput {
  /** @minLength 1 */
  nombre: string;
  dir?: string;
}

export interface ClienteUpdate {
  nombre?: string;
  dir?: string;
}

export type PedidoItemTipo =
  (typeof PedidoItemTipo)[keyof typeof PedidoItemTipo];

export const PedidoItemTipo = {
  garrafa: "garrafa",
  gas: "gas",
} as const;

export interface PedidoItem {
  tipo: PedidoItemTipo;
  prod: string;
  /** @minimum 1 */
  cant: number;
}

export type PedidoTurno = (typeof PedidoTurno)[keyof typeof PedidoTurno];

export const PedidoTurno = {
  manana: "manana",
  tarde: "tarde",
} as const;

/**
 * @nullable
 */
export type PedidoGarrafaEstado =
  | (typeof PedidoGarrafaEstado)[keyof typeof PedidoGarrafaEstado]
  | null;

export const PedidoGarrafaEstado = {
  pendiente: "pendiente",
  paga: "paga",
} as const;

export interface Pedido {
  id: number;
  nombre: string;
  /** @nullable */
  dir: string | null;
  rep: string;
  turno: PedidoTurno;
  tienePedido: boolean;
  /** @nullable */
  nota?: string | null;
  tieneGarrafa: boolean;
  /** @nullable */
  garrafaEstado?: PedidoGarrafaEstado;
  items: PedidoItem[];
  fechaOrigen: string;
  fechaActual: string;
}

export type PedidoInputTurno =
  (typeof PedidoInputTurno)[keyof typeof PedidoInputTurno];

export const PedidoInputTurno = {
  manana: "manana",
  tarde: "tarde",
} as const;

export type PedidoInputGarrafaEstado =
  (typeof PedidoInputGarrafaEstado)[keyof typeof PedidoInputGarrafaEstado];

export const PedidoInputGarrafaEstado = {
  pendiente: "pendiente",
  paga: "paga",
} as const;

export interface PedidoInput {
  nombre: string;
  dir?: string;
  rep: string;
  turno: PedidoInputTurno;
  tienePedido?: boolean;
  nota?: string;
  tieneGarrafa?: boolean;
  garrafaEstado?: PedidoInputGarrafaEstado;
  items: PedidoItem[];
  fechaOrigen: string;
  fechaActual: string;
}

export type PedidoUpdateTurno =
  (typeof PedidoUpdateTurno)[keyof typeof PedidoUpdateTurno];

export const PedidoUpdateTurno = {
  manana: "manana",
  tarde: "tarde",
} as const;

export type PedidoUpdateGarrafaEstado =
  (typeof PedidoUpdateGarrafaEstado)[keyof typeof PedidoUpdateGarrafaEstado];

export const PedidoUpdateGarrafaEstado = {
  pendiente: "pendiente",
  paga: "paga",
} as const;

export interface PedidoUpdate {
  turno?: PedidoUpdateTurno;
  fechaActual?: string;
  garrafaEstado?: PedidoUpdateGarrafaEstado;
  nota?: string;
  tienePedido?: boolean;
}

export interface TodayStats {
  totalManana: number;
  totalTarde: number;
  totalPedidos: number;
  garrafasPendientes: number;
  garrafasPagas: number;
}

export interface HistorialDia {
  fecha: string;
  pedidos: Pedido[];
}

export type ListPedidosParams = {
  /**
   * Filter by fechaActual (YYYY-MM-DD). Omit to get all.
   */
  fecha?: string;
};