# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hoja de Ruta - Gas/Garrafa delivery route management API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: clientes
    description: Customer management
  - name: pedidos
    description: Delivery order management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /clientes:
    get:
      operationId: listClientes
      tags: [clientes]
      summary: List all clients
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cliente"
    post:
      operationId: createCliente
      tags: [clientes]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClienteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
        "409":
          description: Client already exists

  /clientes/{id}:
    patch:
      operationId: updateCliente
      tags: [clientes]
      summary: Update a client's address
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClienteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
        "404":
          description: Not found
    delete:
      operationId: deleteCliente
      tags: [clientes]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /pedidos:
    get:
      operationId: listPedidos
      tags: [pedidos]
      summary: List delivery orders
      parameters:
        - name: fecha
          in: query
          required: false
          schema:
            type: string
          description: Filter by fechaActual (YYYY-MM-DD). Omit to get all.
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pedido"
    post:
      operationId: createPedido
      tags: [pedidos]
      summary: Create a delivery order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PedidoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pedido"

  /pedidos/{id}:
    patch:
      operationId: updatePedido
      tags: [pedidos]
      summary: Update a delivery order (move shifts, update status)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PedidoUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pedido"
        "404":
          description: Not found
    delete:
      operationId: deletePedido
      tags: [pedidos]
      summary: Delete a delivery order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /pedidos/stats/today:
    get:
      operationId: getTodayStats
      tags: [pedidos]
      summary: Get today's delivery stats (totals per shift, repartidor, product type)
      responses:
        "200":
          description: Today stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodayStats"

  /pedidos/historial:
    get:
      operationId: getHistorial
      tags: [pedidos]
      summary: Get order history grouped by fechaOrigen
      responses:
        "200":
          description: History
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HistorialDia"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Cliente:
      type: object
      required: [id, nombre, dir]
      properties:
        id:
          type: integer
        nombre:
          type: string
        dir:
          type: ["string", "null"]

    ClienteInput:
      type: object
      required: [nombre]
      properties:
        nombre:
          type: string
          minLength: 1
        dir:
          type: string

    ClienteUpdate:
      type: object
      properties:
        nombre:
          type: string
        dir:
          type: string

    PedidoItem:
      type: object
      required: [tipo, prod, cant]
      properties:
        tipo:
          type: string
          enum: [garrafa, gas]
        prod:
          type: string
        cant:
          type: integer
          minimum: 1

    Pedido:
      type: object
      required: [id, nombre, dir, rep, turno, tienePedido, tieneGarrafa, items, fechaOrigen, fechaActual]
      properties:
        id:
          type: integer
        nombre:
          type: string
        dir:
          type: ["string", "null"]
        rep:
          type: string
        turno:
          type: string
          enum: [manana, tarde]
        tienePedido:
          type: boolean
        nota:
          type: ["string", "null"]
        tieneGarrafa:
          type: boolean
        garrafaEstado:
          type: ["string", "null"]
          enum: [pendiente, paga, null]
        items:
          type: array
          items:
            $ref: "#/components/schemas/PedidoItem"
        fechaOrigen:
          type: string
        fechaActual:
          type: string

    PedidoInput:
      type: object
      required: [nombre, rep, turno, items, fechaOrigen, fechaActual]
      properties:
        nombre:
          type: string
        dir:
          type: string
        rep:
          type: string
        turno:
          type: string
          enum: [manana, tarde]
        tienePedido:
          type: boolean
        nota:
          type: string
        tieneGarrafa:
          type: boolean
        garrafaEstado:
          type: string
          enum: [pendiente, paga]
        items:
          type: array
          items:
            $ref: "#/components/schemas/PedidoItem"
        fechaOrigen:
          type: string
        fechaActual:
          type: string

    PedidoUpdate:
      type: object
      properties:
        turno:
          type: string
          enum: [manana, tarde]
        fechaActual:
          type: string
        garrafaEstado:
          type: string
          enum: [pendiente, paga]
        nota:
          type: string
        tienePedido:
          type: boolean

    TodayStats:
      type: object
      required: [totalManana, totalTarde, totalPedidos, garrafasPendientes, garrafasPagas]
      properties:
        totalManana:
          type: integer
        totalTarde:
          type: integer
        totalPedidos:
          type: integer
        garrafasPendientes:
          type: integer
        garrafasPagas:
          type: integer

    HistorialDia:
      type: object
      required: [fecha, pedidos]
      properties:
        fecha:
          type: string
        pedidos:
          type: array
          items:
            $ref: "#/components/schemas/Pedido"