# lib/api-zod/src/generated/types/clienteUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hoja de Ruta - Gas/Garrafa delivery route management API
 * OpenAPI spec version: 0.1.0
 */

export interface ClienteUpdate {
  nombre?: string;
  dir?: string;
}