# lib/api-zod/src/generated/types/historialDia.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hoja de Ruta - Gas/Garrafa delivery route management API
 * OpenAPI spec version: 0.1.0
 */
import type { Pedido } from "./pedido";

export interface HistorialDia {
  fecha: string;
  pedidos: Pedido[];
}