# lib/api-zod/src/generated/types/pedido.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hoja de Ruta - Gas/Garrafa delivery route management API
 * OpenAPI spec version: 0.1.0
 */
import type { PedidoGarrafaEstado } from "./pedidoGarrafaEstado";
import type { PedidoItem } from "./pedidoItem";
import type { PedidoTurno } from "./pedidoTurno";

export interface Pedido {
  id: number;
  nombre: string;
  /** @nullable */
  dir: string | null;
  rep: string;
  turno: PedidoTurno;
  tienePedido: boolean;
  /** @nullable */
  nota?: string | null;
  tieneGarrafa: boolean;
  /** @nullable */
  garrafaEstado?: PedidoGarrafaEstado;
  items: PedidoItem[];
  fechaOrigen: string;
  fechaActual: string;
}