# lib/api-zod/src/generated/types/pedidoInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hoja de Ruta - Gas/Garrafa delivery route management API
 * OpenAPI spec version: 0.1.0
 */
import type { PedidoInputGarrafaEstado } from "./pedidoInputGarrafaEstado";
import type { PedidoInputTurno } from "./pedidoInputTurno";
import type { PedidoItem } from "./pedidoItem";

export interface PedidoInput {
  nombre: string;
  dir?: string;
  rep: string;
  turno: PedidoInputTurno;
  tienePedido?: boolean;
  nota?: string;
  tieneGarrafa?: boolean;
  garrafaEstado?: PedidoInputGarrafaEstado;
  items: PedidoItem[];
  fechaOrigen: string;
  fechaActual: string;
}