# lib/api-zod/src/generated/types/pedidoItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hoja de Ruta - Gas/Garrafa delivery route management API
 * OpenAPI spec version: 0.1.0
 */
import type { PedidoItemTipo } from "./pedidoItemTipo";

export interface PedidoItem {
  tipo: PedidoItemTipo;
  prod: string;
  /** @minimum 1 */
  cant: number;
}