# lib/api-zod/src/generated/types/pedidoUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hoja de Ruta - Gas/Garrafa delivery route management API
 * OpenAPI spec version: 0.1.0
 */
import type { PedidoUpdateGarrafaEstado } from "./pedidoUpdateGarrafaEstado";
import type { PedidoUpdateTurno } from "./pedidoUpdateTurno";

export interface PedidoUpdate {
  turno?: PedidoUpdateTurno;
  fechaActual?: string;
  garrafaEstado?: PedidoUpdateGarrafaEstado;
  nota?: string;
  tienePedido?: boolean;
}